third_party/blink/web_tests/http/tests/devtools/elements/shadow/create-shadow-root-expected.txt

Tests that elements panel updates dom tree structure upon shadow root creation.


Running: testCreateShadowRoot
==== before ====
- <div id="container">
      <div id="child"></div>
  </div>
==== after ====
- <div id="container">
    - #shadow-root (open)
          <div id="shadow-1"></div>
      <div id="child"></div>
  </div>

Running: testCreateOpenShadowRoot
==== before ====
- <div id="containerOpen">
      <div id="childOpen"></div>
  </div>
==== after ====
- <div id="containerOpen">
    - #shadow-root (open)
          <div id="shadow-3"></div>
      <div id="childOpen"></div>
  </div>

Running: testCreateCloseShadowRoot
==== before ====
- <div id="containerClosed">
      <div id="childClosed"></div>
  </div>
==== after ====
- <div id="containerClosed">
    - #shadow-root (closed)
          <div id="shadow-4"></div>
      <div id="childClosed"></div>
  </div>
